% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_beta_diversity.R
\name{plot_beta_diversity}
\alias{plot_beta_diversity}
\title{Analysis and plot of Beta diversity (Turnover)}
\usage{
plot_beta_diversity()
}
\description{
Analysis and plot of Beta diversity (Turnover)
}
