% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_gymno_shannon.R, R/soil_to_sr.R
\name{soil_to_sr}
\alias{soil_to_sr}
\title{analyse and plot shannon diversity as a function of gymnosperm concentration}
\usage{
soil_to_sr()

soil_to_sr()
}
\description{
analyse and plot shannon diversity as a function of gymnosperm concentration

analyse and plot shannon diversity as a function of gymnosperm concentration
}
