% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats_for_soil_properties.R
\name{stats_for_soil_properties}
\alias{stats_for_soil_properties}
\title{Run linear mixed effect models on the soil properties as a function of windbreaks}
\usage{
stats_for_soil_properties(y = main_data$Whc)
}
\arguments{
\item{y}{a vector of data from the main_data dataset.}
}
\value{
list of relevant stats.
}
\description{
This function statistically tests the soil property given for example:soil water content,
water holdin capacity, organic C, total N, C:N ratio, and soil pH and regress it to the distance
from the windbreak while having fields as a random effect.
}
\details{
No details for now.
}
\examples{
stats <- stats_for_soil_properties(y = main_data$Whc)
}
\seealso{
\code{\link{agroPack}}.
}
\author{
Maxime Rivest
}
\keyword{manip}
