% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_of_multiple_comparaison.R
\name{table_of_multiple_comparaison}
\alias{table_of_multiple_comparaison}
\title{Present the results of a mixed effect models on the soil properties as a function of windbreaks}
\usage{
table_of_multiple_comparaison(my_list, varname = "Name to display")
}
\arguments{
\item{my_list}{a list out of the function stats_for_soil_properties}

\item{varname}{a character vector of length one holding the name of the variable that should
be displayed in the table}
}
\value{
data.frame of coefficients for multiple comparaisons
}
\description{
Generate a clean data frame.
}
\details{
No details for now.
}
\examples{
stats <- stats_for_soil_properties(y = main_data$Whc)
table_of_multiple_comparaison(my_list = stats, varname = "Water holding capacity")
}
\seealso{
\code{\link{agroPack}}.
}
\author{
Maxime Rivest
}
\keyword{manip}
