% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_Publication.R
\name{theme_Publication}
\alias{theme_Publication}
\title{Ggplot theme for publication}
\usage{
theme_Publication()
}
\arguments{
\item{y}{a vector of data from the main_data dataset.}
}
\value{
list of relevant stats.
}
\description{
A theme
}
\details{
No details for now.
}
\examples{
stats <- stats_for_soil_properties(y = main_data$Whc)
}
\seealso{
\code{\link{agroPack}}.
}
\author{
Maxime Rivest
}
\keyword{manip}
